python_engine/face_analysis/requirements.txt
# Fashion Expert - ML/Computer Vision dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6

# ML/DL Stack
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
mediapipe>=0.10.0
opencv-python-headless>=4.8.0
pillow>=10.0.0
scikit-image>=0.21.0
scikit-learn>=1.3.0
numpy>=1.24.0

# Utilities
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
